backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
mysqlclient
django-cors-headers